Migrate orderly to outpack

Migrate an orderly archive to outpack. This function is subject to change, and does not handle all failure modes. It is quite slow because it involves hashing the contents of your archive several times - if you have GBs of files expect this to take minutes or hours.

orderly2outpack(src, dest)

Arguments

src

Path to the orderly archive to migrate. This must be "complete" (i.e., containing a full graph and no metadata only pulls)

dest

The destination to create a new outpack archive. This will be created with no user-visible archive directory, with a file store, and requiring a full tree (i.e., in a suitable configuration to use as an upstream source).

Value

The path to the newly created archive