Migrate orderly source

Migrate an orderly source directory for use with outpack. This works in place, writing out <name>/<name>.R files and deleting orderly.yml files (optionally). This is intended to be run on a clean clone of your source git repository.

orderly2outpack_src(path, delete_yml = FALSE, strict = FALSE, dry_run = FALSE)

Arguments

path

Path to the orderly source directory

delete_yml

Logical, indicating if the orderly.yml files should be deleted.

strict

Logical, indicating if we should enable orderly2::orderly_strict_mode() in the resulting source files. You can always add this later to specific cases (or remove it).

dry_run

Logical, indicating if we should just run process all the data, but not write any changes.

Value

Nothing, called for side effects only